Migrating from `monit` to `supervisor`.
This commit references #22 and establishes the first attempt to migrate
from `monit` to `supervisor`. The current code doesn't address the fact
that we can't start `zabbix-*` services without daemonizaing them. This
is required since `supervisor` must spawn such processes and hold their
reference and therefore the ability to assess if they are up or not.